[Gold V] Title: 옥상 정원 꾸미기, Time: 968 ms, Memory: 21660 KB -BaekjoonHub

+++ 백준/Gold/6198. 옥상 정원 꾸미기/옥상 정원 꾸미기.java
@@ -7,11 +7,11 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         int N = Integer.parseInt(br.readLine());
-        long[] apt = new long[N];
+        int[] apt = new int[N];
         long total = 0l;
 
         for (int i = 0; i < N; i++) {
-            apt[i] = Long.parseLong(br.readLine());
+            apt[i] = Integer.parseInt(br.readLine());
         }
 
         for (int i = 0; i < N - 1; i++) {
